+++ packages/components/src/Dialog/Layout/DialogContent.test.tsx
@@ -24,43 +24,19 @@
 
  */
 
-export * from './Accordion'
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './MessageBar'
-export * from './Dialog'
-export * from './PageSize'
-export * from './Pagination'
-export * from './Popover'
-export * from './Spinner'
-export * from './Status'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './Tree'
-export * from './VisuallyHidden'
+import 'jest-styled-components'
+import React from 'react'
+import { assertSnapshot } from '@looker/components-test-utils'
+import { DialogContent } from './DialogContent'
 
-export * from './utils'
+test('DialogContent - no overflow', () => {
+  assertSnapshot(<DialogContent>Stuff</DialogContent>)
+})
 
-export { ComponentsProvider } from '@looker/components-providers'
-
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
-
-export { SpacingSizes, theme, Theme } from '@looker/design-tokens'
+test('DialogContent - no overflow', () => {
+  assertSnapshot(
+    <DialogContent>
+      <div style={{ height: '4rem' }}>Stuff</div>
+    </DialogContent>
+  )
+})