+++ packages/components/src/Dialog/Layout/DialogFooter.test.tsx
@@ -24,43 +24,36 @@
 
  */
 
-export * from './Accordion'
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './MessageBar'
-export * from './Dialog'
-export * from './PageSize'
-export * from './Pagination'
-export * from './Popover'
-export * from './Spinner'
-export * from './Status'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './Tree'
-export * from './VisuallyHidden'
+import 'jest-styled-components'
+import React from 'react'
+import { assertSnapshot } from '@looker/components-test-utils'
+import { DialogContext } from '../DialogContext'
+import { DialogFooter } from './DialogFooter'
 
-export * from './utils'
+test('DialogFooter with Button', () => {
+  assertSnapshot(
+    <DialogFooter>
+      <button>Cancel</button>
+    </DialogFooter>
+  )
+})
 
-export { ComponentsProvider } from '@looker/components-providers'
-
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
-
-export { SpacingSizes, theme, Theme } from '@looker/design-tokens'
+test('DialogFooter with DialogContext', () => {
+  assertSnapshot(
+    <DialogContext.Consumer>
+      {({ closeModal }) => (
+        <DialogFooter>
+          <button onClick={closeModal}>Cancel</button>
+          <button
+            onClick={() => {
+              alert('Doing things...')
+              closeModal && closeModal()
+            }}
+          >
+            Yes, Delete "Stuff"
+          </button>
+        </DialogFooter>
+      )}
+    </DialogContext.Consumer>
+  )
+})