+++ packages/components/src/Dialog/Layout/DialogHeader.test.tsx
@@ -24,43 +24,42 @@
 
  */
 
-export * from './Accordion'
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './MessageBar'
-export * from './Dialog'
-export * from './PageSize'
-export * from './Pagination'
-export * from './Popover'
-export * from './Spinner'
-export * from './Status'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './Tree'
-export * from './VisuallyHidden'
+import 'jest-styled-components'
+import React from 'react'
+import {
+  assertSnapshot,
+  mountWithTheme,
+  renderWithTheme,
+} from '@looker/components-test-utils'
+import { IconButton } from '../../Button'
+import { DialogHeader } from './DialogHeader'
 
-export * from './utils'
+test('DialogHeader', () => {
+  assertSnapshot(
+    <DialogHeader id="test-DialogHeader">The Heading for a Dialog</DialogHeader>
+  )
+})
 
-export { ComponentsProvider } from '@looker/components-providers'
+test('DialogHeader passes through DOM props', () => {
+  const { findByLabelText } = renderWithTheme(
+    <DialogHeader aria-label="This is the ARIA label">
+      The Heading for a Dialog
+    </DialogHeader>
+  )
 
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
+  expect(findByLabelText('This is the ARIA label')).toBeTruthy()
+})
 
-export { SpacingSizes, theme, Theme } from '@looker/design-tokens'
+test('DialogHeader with hideClose', () => {
+  const withClose = mountWithTheme(
+    <DialogHeader id="test-DialogHeader">The Heading for a Dialog</DialogHeader>
+  )
+  expect(withClose.find(IconButton).exists()).toBeTruthy()
+
+  const withoutClose = mountWithTheme(
+    <DialogHeader id="test-DialogHeader" hideClose>
+      The Heading for a Dialog
+    </DialogHeader>
+  )
+  expect(withoutClose.find(IconButton).exists()).toBeFalsy()
+})