+++ packages/components/src/Popover/Popover.tsx
@@ -36,7 +36,8 @@ import React, {
   SyntheticEvent,
 } from 'react'
 import { Box } from '../Layout'
-import { ModalPortal, ModalContext } from '../Modal'
+import { Portal } from '../Portal'
+import { DialogContext } from '../Dialog'
 import { OverlaySurface } from '../Overlay/OverlaySurface'
 import {
   useCallbackRef,
@@ -385,7 +386,7 @@ export function usePopover({
     trapRef: focusTrapRef,
   } = useFocusTrap(controlledIsOpen && focusTrap)
 
-  const { focusTrapRef: parentFocusTrapRef } = useContext(ModalContext)
+  const { focusTrapRef: parentFocusTrapRef } = useContext(DialogContext)
 
   const [newTriggerElement, callbackRef] = useCallbackRef()
   // If the triggerElement is passed in props, use that instead of the new element
@@ -474,7 +475,7 @@ export function usePopover({
   const [containerElement, contentContainerRef] = useCallbackRef<HTMLElement>()
 
   const popover = !openWithoutElem && isOpen && (
-    <ModalContext.Provider
+    <DialogContext.Provider
       value={{
         closeModal: handleClose,
         disableFocusTrap,
@@ -486,7 +487,7 @@ export function usePopover({
         scrollLockEnabled,
       }}
     >
-      <ModalPortal ref={scrollRef}>
+      <Portal ref={scrollRef}>
         <OverlaySurface
           arrow={arrow}
           arrowProps={arrowProps}
@@ -509,8 +510,8 @@ export function usePopover({
             {content}
           </Box>
         </OverlaySurface>
-      </ModalPortal>
-    </ModalContext.Provider>
+      </Portal>
+    </DialogContext.Provider>
   )
   return {
     contentContainer: containerElement,