+++ www/src/MDX/Pre/allComponents.ts
@@ -61,10 +61,13 @@ import {
   DateTimeFormat,
   doDefaultActionListSort,
   Dialog,
+  DialogContent,
+  DialogContext,
+  DialogFooter,
+  DialogHeader,
   DialogManager,
+  DialogSurface,
   Divider,
-  Drawer,
-  DrawerManager,
   FieldCheckbox,
   FieldCheckboxGroup,
   FieldChips,
@@ -113,11 +116,6 @@ import {
   MenuItem,
   MenuList,
   MessageBar,
-  ModalContent,
-  ModalContext,
-  ModalFooter,
-  ModalHeader,
-  ModalSurface,
   PageSize,
   Pagination,
   Paragraph,
@@ -162,11 +160,9 @@ import {
   useToggle,
   VisuallyHidden,
 } from '@looker/components'
-
 import styled from 'styled-components'
 import isEqual from 'lodash/isEqual'
 import { useState } from 'react'
-
 import { GridPlaceholder } from '../../helpers/GridPlaceholder'
 
 const otherLibraries = {
@@ -224,10 +220,13 @@ export const allComponents = {
   DateFormat,
   DateTimeFormat,
   Dialog,
+  DialogContent,
+  DialogContext,
+  DialogFooter,
+  DialogHeader,
   DialogManager,
+  DialogSurface,
   Divider,
-  Drawer,
-  DrawerManager,
   FieldCheckbox,
   FieldCheckboxGroup,
   FieldChips,
@@ -276,11 +275,6 @@ export const allComponents = {
   MenuItem,
   MenuList,
   MessageBar,
-  ModalContent,
-  ModalContext,
-  ModalFooter,
-  ModalHeader,
-  ModalSurface,
   PageSize,
   Pagination,
   Paragraph,